Restructure - Take III - Follow some package best practices.
Create two main folders:
- 'scripts', executable scripts
- 'snowrs', shared code used by the scripts
- Merged all script helper methods to one module
