+++ .travis.yml
@@ -11,6 +11,7 @@ install:
   - conda env create -q --file $TRAVIS_BUILD_DIR/conda_environment.txt -p $HOME/miniconda/envs/snow_rs
 
 before_script:
+  - export PYTHONPATH=$TRAVIS_BUILD_DIR
   - conda env list
   - source activate snow_rs
   - cd $TRAVIS_BUILD_DIR/
