--- modis/__init__.py
@@ -1,6 +1,0 @@
-from .script_helpers import parse_year, validate_types
-
-__all__ = [
-    'parse_year',
-    'validate_types'
-]
