+++ snowrs/script_helpers.py
@@ -1,13 +1,25 @@
 import os
-
 from datetime import date, datetime, timedelta
 
+from snowrs import SourceFolder
 
-def ensure_folder(folder):
-    if not (os.path.isdir(folder)):
-        os.mkdir(folder)
 
-    return folder
+def validate_types(ctx, _param, value):
+    if not SourceFolder.valid_source_type(value):
+        print('Invalid source data type\n' +
+              'Possible options:\n' +
+              ' * forcing\n' +
+              ' * fraction')
+        ctx.abort()
+    else:
+        return value
+
+
+def parse_year(_ctx, _param, value):
+    if value:
+        return range(value, value + 1)
+    else:
+        return range(2000, date.today().year + 1)
 
 
 def dates_in_range(year, from_day, to_day):
